requirements.txt
# Core LLM & Vector DB
qdrant-client>=1.7.0
aiohttp>=3.9.0
httpx>=0.24.0
litellm
tenacity
tavily-python

# Document processing
pypdf>=3.18.0
python-docx>=0.8.11
markdown>=3.5.0
python-multipart

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async & concurrency
asyncio>=3.4.3
aiofiles>=23.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Development & testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
